fix(onboarding): combines ghgi inventory uploads flow with city details steps
