+++ app/src/components/steps/GHGI/import/ghgi-import-wizard.tsx
@@ -1,0 +1,1073 @@
+"use client";
+
+import { useTranslation } from "@/i18n/client";
+import { MdArrowBack, MdArrowForward, MdWarning } from "react-icons/md";
+import { Box, HStack, Icon, Text, useSteps } from "@chakra-ui/react";
+import { useRouter, usePathname } from "next/navigation";
+import React, { useState, useEffect, useRef, useCallback, useMemo } from "react";
+import { motion, AnimatePresence } from "framer-motion";
+import ProgressSteps from "@/components/steps/progress-steps";
+import { Button } from "@/components/ui/button";
+import { UseErrorToast, UseInfoToast, UseSuccessToast } from "@/hooks/Toasts";
+import UploadFileStep from "@/components/steps/GHGI/import/upload-file-step";
+import InventoryMappingStep from "@/components/steps/GHGI/import/inventory-mapping-step";
+import ReviewConfirmStep from "@/components/steps/GHGI/import/review-confirm-step";
+import ImportingInventoryStep from "@/components/steps/GHGI/import/importing-inventory-step";
+import DataLossWarningModal from "@/components/Modals/data-loss-warning-modal";
+import { api } from "@/services/api";
+import { logger } from "@/services/logger";
+import { getApiErrorMessage } from "@/util/helpers";
+import type { ImportStatusResponse } from "@/util/types";
+import { readImportChunkProgress } from "@/util/import-chunk-progress";
+import { usePollUntil } from "@/hooks/usePollUntil";
+import { TFunction } from "i18next";
+import { useAppDispatch } from "@/lib/hooks";
+
+/** If errorLog starts with "i18n:", return t(key); else return errorLog or fallback. */
+function resolveErrorMessage(
+  errorLog: string | null | undefined,
+  fallback: string,
+  t: TFunction,
+): string {
+  if (errorLog == null || errorLog === "") return fallback;
+  if (errorLog.startsWith("i18n:")) return t(errorLog.slice(5));
+  return errorLog;
+}
+
+function ImportButton({
+  cityId,
+  inventoryId,
+  importedFileId,
+  mappingOverrides,
+  onImport,
+  onImportProgressChange,
+  t,
+}: {
+  cityId: string;
+  inventoryId: string;
+  importedFileId: string;
+  mappingOverrides: Record<string, string>;
+  onImport: () => void;
+  onImportProgressChange: (inProgress: boolean) => void;
+  t: TFunction;
+}) {
+  const dispatch = useAppDispatch();
+  const [approveImport, { isLoading: isImporting }] =
+    api.useApproveImportMutation();
+  const [getImportStatus] = api.useLazyGetImportStatusQuery();
+
+  /** Invalidate inventory-related caches so the GHGI page shows fresh totalEmissions. */
+  const invalidateInventoryCache = () => {
+    dispatch(
+      api.util.invalidateTags([
+        "Inventory",
+        "InventoryProgress",
+        "InventoryValue",
+        "ReportResults",
+        "YearlyReportResults",
+      ]),
+    );
+  };
+
+  const makeErrorToast = (title: string, description?: string) => {
+    const { showErrorToast } = UseErrorToast({ description, title });
+    showErrorToast();
+  };
+
+  const makeSuccessToast = (title: string, description?: string) => {
+    const { showSuccessToast } = UseSuccessToast({ description, title });
+    showSuccessToast();
+  };
+
+  const { startPolling: startImportPolling, stopPolling: stopImportPolling, isPolling: isImportPolling } =
+    usePollUntil<ImportStatusResponse>({
+      fetch: () =>
+        getImportStatus({ cityId, inventoryId, importedFileId }).unwrap() as Promise<ImportStatusResponse>,
+      isTerminal: (res) => {
+        if (res.importStatus === "completed")
+          return { done: true, success: true, data: res };
+        if (res.importStatus === "failed")
+          return { done: true, success: false, data: res };
+        return { done: false };
+      },
+      onSuccess: () => {
+        invalidateInventoryCache();
+        makeSuccessToast(t("import-completed"), t("import-completed-description"));
+        onImport();
+      },
+      onFailure: (res) => {
+        onImportProgressChange(false);
+        makeErrorToast("Import failed", resolveErrorMessage(res.errorLog, "Failed to import data", t));
+      },
+      onPollError: (err) =>
+        logger.debug(
+          { err, cityId, inventoryId, importedFileId },
+          "Approve/import status poll failed",
+        ),
+      intervalMs: 3000,
+    });
+
+  const handleImport = async () => {
+    if (!importedFileId) return;
+    stopImportPolling();
+    // Swap ReviewConfirmStep for the importing panel before approve polling
+    // can clear reviewData from the shared status cache.
+    onImportProgressChange(true);
+
+    const overridesToSend = Object.fromEntries(
+      Object.entries(mappingOverrides).filter(([, v]) => v !== ""),
+    );
+
+    try {
+      const result = await approveImport({
+        cityId,
+        inventoryId,
+        importedFileId,
+        mappingOverrides: Object.keys(overridesToSend).length
+          ? overridesToSend
+          : undefined,
+      }).unwrap();
+
+      if ("accepted" in result && result.accepted) {
+        startImportPolling();
+        return;
+      }
+
+      if ((result as { importStatus?: string }).importStatus === "completed") {
+        invalidateInventoryCache();
+        makeSuccessToast(t("import-completed"), t("import-completed-description"));
+        onImport();
+        return;
+      }
+      if ((result as { importStatus?: string }).importStatus === "failed") {
+        onImportProgressChange(false);
+        makeErrorToast(
+          "Import failed",
+          resolveErrorMessage(
+            (result as { errorLog?: string | null }).errorLog,
+            "Failed to import data",
+            t,
+          ),
+        );
+        return;
+      }
+      onImportProgressChange(false);
+    } catch (error: unknown) {
+      onImportProgressChange(false);
+      makeErrorToast(
+        "Import failed",
+        getApiErrorMessage(error, "Failed to import data"),
+      );
+    }
+  };
+
+  return (
+    <Button
+      w="auto"
+      gap="8px"
+      py="16px"
+      px="24px"
+      onClick={handleImport}
+      h="64px"
+      loading={isImporting || isImportPolling}
+      disabled={isImporting || isImportPolling}
+    >
+      <Text
+        fontFamily="button.md"
+        fontWeight="600"
+        letterSpacing="wider"
+      >
+        {t("import-inventory")}
+      </Text>
+      <MdArrowForward height="24px" width="24px" />
+    </Button>
+  );
+}
+
+export type GhgiImportWizardProps = {
+  lng: string;
+  cityId: string;
+  inventoryId: string;
+  /** Called when import completes successfully. */
+  onComplete: (inventoryId: string) => void;
+  /**
+   * Full ProgressSteps list when this wizard is embedded in a longer flow
+   * (e.g. GHGI setup details + population + these 3 import steps).
+   * Defaults to the 3 import-only titles.
+   */
+  progressSteps?: { title: string }[];
+  /** Index offset into `progressSteps` for the first import step (default 0). */
+  progressStepOffset?: number;
+  /** Back from the first import step when embedded (instead of router.back). */
+  onExitFirstStep?: () => void;
+};
+
+/**
+ * Upload → mapping → review import wizard.
+ * Used standalone on `/import` and embedded in GHGI setup upload mode (5-step stepper).
+ */
+export default function GhgiImportWizard({
+  lng,
+  cityId,
+  inventoryId,
+  onComplete,
+  progressSteps,
+  progressStepOffset = 0,
+  onExitFirstStep,
+}: GhgiImportWizardProps) {
+  const { t } = useTranslation(lng, "onboarding");
+  const router = useRouter();
+
+  const defaultSteps = [
+    { title: t("upload-file-step") },
+    { title: t("inventory-mapping-step") },
+    { title: t("review-confirm-step") },
+  ];
+  const steps = progressSteps ?? defaultSteps;
+  const importStepCount = defaultSteps.length;
+
+  const {
+    value: activeStep,
+    goToNextStep,
+    goToPrevStep,
+    setStep,
+  } = useSteps({
+    defaultStep: 0,
+    count: importStepCount,
+  });
+
+  // Keep the combined stepper in sync with the absolute index.
+  const progressCurrentStep = progressStepOffset + activeStep;
+
+  const [uploadedFile, setUploadedFile] = useState<File | null>(null);
+  const [importedFileId, setImportedFileId] = useState<string | null>(null);
+  const [lastImportStatus, setLastImportStatus] = useState<ImportStatusResponse | null>(null);
+  const [pdfPendingExtraction, setPdfPendingExtraction] = useState(false);
+  const [tabularPendingInterpretation, setTabularPendingInterpretation] =
+    useState(false);
+  const [isExtractInProgress, setIsExtractInProgress] = useState(false);
+  const [isInterpretInProgress, setIsInterpretInProgress] = useState(false);
+  const [isImportInProgress, setIsImportInProgress] = useState(false);
+  const [mappingOverrides, setMappingOverrides] = useState<
+    Record<string, string>
+  >({});
+  const [showDataLossModal, setShowDataLossModal] = useState(false);
+  const [pendingNavigation, setPendingNavigation] = useState<(() => void) | null>(null);
+  const [extractionProgress, setExtractionProgress] = useState<{
+    current: number;
+    total: number;
+  } | null>(null);
+  const uploadPendingFileRef = useRef<File | null>(null);
+  const uploadPendingIdRef = useRef<string | null>(null);
+  const pathname = usePathname();
+  const prevPathnameRef = useRef<string | null>(null);
+  const fileYearMismatchToastShownRef = useRef(false);
+
+  // Check if there's unsaved progress
+  const hasUnsavedProgress = uploadedFile !== null || importedFileId !== null;
+
+  const makeErrorToast = (title: string, description?: string) => {
+    const { showErrorToast } = UseErrorToast({ description, title });
+    showErrorToast();
+  };
+
+  const makeInfoToast = (title: string, description?: string) => {
+    const { showInfoToast } = UseInfoToast({ description, title });
+    showInfoToast();
+  };
+
+  const [uploadFile, { isLoading: isUploadingFile }] =
+    api.useUploadInventoryFileMutation();
+  const [extractImport, { isLoading: isExtracting }] =
+    api.useExtractImportMutation();
+  const [interpretImport, { isLoading: isInterpreting }] =
+    api.useInterpretImportMutation();
+  const [getImportStatus] = api.useLazyGetImportStatusQuery();
+  const { data: mappingStepData } = api.useGetImportStatusQuery(
+    { cityId, inventoryId: inventoryId ?? "", importedFileId: importedFileId ?? "" },
+    { skip: !importedFileId || !inventoryId },
+  );
+  const { data: inventory } = api.useGetInventoryQuery(inventoryId ?? "", {
+    skip: !inventoryId,
+  });
+  const { data: inventoryProgress } = api.useGetInventoryProgressQuery(
+    inventoryId ?? "",
+    { skip: !inventoryId },
+  );
+
+  // Check if the inventory already contains data (any InventoryValue records)
+  const inventoryHasData = useMemo(() => {
+    if (!inventoryProgress?.totalProgress) return false;
+    const { thirdParty, uploaded, reasonNE, reasonNO } =
+      inventoryProgress.totalProgress;
+    return thirdParty + uploaded + reasonNE + reasonNO > 0;
+  }, [inventoryProgress]);
+
+  const canContinueMapping = useMemo(() => {
+    const cols = mappingStepData?.columnMappings?.columns ?? [];
+    const reqMappings = mappingStepData?.columnMappings?.requiredMappings ?? [];
+    if (cols.length === 0 || reqMappings.length === 0) return false;
+    const MANDATORY = new Set(["gpcRefNo", "sector", "subsector", "activityAmount"]);
+    const keyForLabel = (label: string | null) =>
+      label ? (reqMappings.find((r) => r.label === label)?.key ?? "") : "";
+    const effectiveKey = (col: { columnName: string; interpretedAs: string | null }) =>
+      col.columnName in mappingOverrides
+        ? mappingOverrides[col.columnName]
+        : keyForLabel(col.interpretedAs);
+    const isMandatory = (col: { columnName: string; interpretedAs: string | null }) =>
+      MANDATORY.has(keyForLabel(col.interpretedAs)) ||
+      MANDATORY.has(mappingOverrides[col.columnName] ?? "");
+    return cols.filter(isMandatory).every((col) => effectiveKey(col) !== "");
+  }, [mappingStepData, mappingOverrides]);
+
+  // Reset year-mismatch toast when the user switches to a different import
+  useEffect(() => {
+    fileYearMismatchToastShownRef.current = false;
+  }, [importedFileId]);
+
+  const inventoryYear =
+    inventory?.year != null && Number.isFinite(Number(inventory.year))
+      ? Number(inventory.year)
+      : null;
+  const fileYear =
+    lastImportStatus?.inferredYearFromFile != null &&
+      Number.isFinite(Number(lastImportStatus.inferredYearFromFile))
+      ? Number(lastImportStatus.inferredYearFromFile)
+      : null;
+  const fileYearMismatch =
+    inventoryYear != null && fileYear != null && inventoryYear !== fileYear;
+
+  useEffect(() => {
+    if (!fileYearMismatch || fileYearMismatchToastShownRef.current) return;
+    fileYearMismatchToastShownRef.current = true;
+    makeErrorToast(t("file-year-mismatch-title"), t("file-year-mismatch", { year: inventoryYear }));
+  }, [fileYearMismatch, inventoryYear, t]);
+
+  const {
+    startPolling: startUploadPolling,
+    stopPolling: stopUploadPolling,
+    isPolling: isUploadPolling,
+  } = usePollUntil<ImportStatusResponse>({
+    fetch: useCallback(() => {
+      const id = uploadPendingIdRef.current;
+      if (!id || !cityId || !inventoryId)
+        return Promise.reject(new Error("Upload poll: missing id or context"));
+      return getImportStatus({ cityId, inventoryId, importedFileId: id }).unwrap() as Promise<ImportStatusResponse>;
+    }, [cityId, inventoryId, getImportStatus]),
+    isTerminal: (res) => {
+      if (
+        res.importStatus === "pending_ai_extraction" ||
+        res.importStatus === "pending_ai_interpretation" ||
+        res.importStatus === "waiting_for_approval"
+      )
+        return { done: true, success: true, data: res };
+      if (res.importStatus === "failed") return { done: true, success: false, data: res };
+      return { done: false };
+    },
+    onSuccess: (res) => {
+      setLastImportStatus(res);
+      const file = uploadPendingFileRef.current;
+      if (file) setUploadedFile(file);
+      setImportedFileId(res.id);
+      if (res.importStatus === "pending_ai_extraction") {
+        setPdfPendingExtraction(true);
+        setTabularPendingInterpretation(false);
+      } else {
+        setPdfPendingExtraction(false);
+        setTabularPendingInterpretation(res.importStatus === "pending_ai_interpretation");
+        if (res.importStatus === "waiting_for_approval") setTimeout(() => goToNextStep(), 150);
+      }
+    },
+    onFailure: (res) =>
+      makeErrorToast("Upload failed", resolveErrorMessage(res.errorLog, "File validation or processing failed", t)),
+    onPollError: (err) =>
+      logger.debug(
+        { err, cityId, inventoryId, importedFileId: uploadPendingIdRef.current },
+        "Upload status poll failed",
+      ),
+    intervalMs: 3000,
+  });
+
+  const {
+    startPolling: startExtractionPolling,
+    stopPolling: stopExtractionPolling,
+  } = usePollUntil<ImportStatusResponse>({
+    fetch: useCallback(() => {
+      if (!importedFileId || !inventoryId) return Promise.reject(new Error("Missing importedFileId or inventoryId"));
+      return getImportStatus({
+        cityId,
+        inventoryId: inventoryId as string,
+        importedFileId: importedFileId as string,
+      }).unwrap() as Promise<ImportStatusResponse>;
+    }, [cityId, inventoryId, importedFileId, getImportStatus]),
+    isTerminal: (res) => {
+      if (res.importStatus === "waiting_for_approval") return { done: true, success: true, data: res };
+      if (res.importStatus === "failed") return { done: true, success: false, data: res };
+      return { done: false };
+    },
+    onSuccess: (res) => {
+      setLastImportStatus(res);
+      setIsExtractInProgress(false);
+      setExtractionProgress(null);
+      setPdfPendingExtraction(false);
+      setTimeout(() => goToNextStep(), 150);
+    },
+    onFailure: (res) => {
+      setIsExtractInProgress(false);
+      setExtractionProgress(null);
+      makeErrorToast(t("extraction-failed"), resolveErrorMessage(res.errorLog, t("ai-extraction-failed-default"), t));
+    },
+    onTick: (res) => {
+      const progress = readImportChunkProgress(res.mappingConfiguration);
+      if (progress) setExtractionProgress(progress);
+    },
+    onPollError: (err) =>
+      logger.debug({ err, cityId, inventoryId, importedFileId }, "Import status poll failed"),
+    intervalMs: 3000,
+  });
+
+  const {
+    startPolling: startInterpretPolling,
+    stopPolling: stopInterpretPolling,
+  } = usePollUntil<ImportStatusResponse>({
+    fetch: useCallback(() => {
+      if (!importedFileId || !inventoryId) return Promise.reject(new Error("Missing importedFileId or inventoryId"));
+      return getImportStatus({
+        cityId,
+        inventoryId: inventoryId as string,
+        importedFileId: importedFileId as string,
+      }).unwrap() as Promise<ImportStatusResponse>;
+    }, [cityId, inventoryId, importedFileId, getImportStatus]),
+    isTerminal: (res) => {
+      if (res.importStatus === "waiting_for_approval") return { done: true, success: true, data: res };
+      if (res.importStatus === "failed") return { done: true, success: false, data: res };
+      return { done: false };
+    },
+    onSuccess: (res) => {
+      setLastImportStatus(res);
+      setIsInterpretInProgress(false);
+      setExtractionProgress(null);
+      setTabularPendingInterpretation(false);
+      setTimeout(() => goToNextStep(), 150);
+    },
+    onFailure: (res) => {
+      setIsInterpretInProgress(false);
+      setExtractionProgress(null);
+      makeErrorToast(
+        t("interpretation-failed") ?? "Interpretation failed",
+        resolveErrorMessage(res.errorLog, t("ai-extraction-failed-default"), t),
+      );
+    },
+    onTick: (res) => {
+      // Path B shape progress (U2) — same payload key as Path C extract
+      const progress = readImportChunkProgress(res.mappingConfiguration);
+      if (progress) setExtractionProgress(progress);
+    },
+    onPollError: (err) =>
+      logger.debug({ err, cityId, inventoryId, importedFileId }, "Interpret status poll failed"),
+    intervalMs: 3000,
+  });
+
+  const handleFileUpload = async (file: File) => {
+    if (!inventoryId) {
+      makeErrorToast("Error", "Inventory ID is required");
+      return;
+    }
+    stopUploadPolling();
+    makeInfoToast(
+      t("upload-started"),
+      t("upload-started-description", { fileName: file.name }),
+    );
+
+    try {
+      const result = await uploadFile({
+        cityId,
+        inventoryId,
+        file,
+      }).unwrap();
+
+      if ("accepted" in result && result.accepted) {
+        uploadPendingFileRef.current = file;
+        uploadPendingIdRef.current = result.id;
+        startUploadPolling();
+        return;
+      }
+
+      setUploadedFile(file);
+      setImportedFileId(result.id);
+      setPdfPendingExtraction(
+        (result as { importStatus?: string; fileType?: string }).importStatus === "pending_ai_extraction" ||
+        (result as { fileType?: string }).fileType === "pdf",
+      );
+      setTabularPendingInterpretation(
+        (result as { importStatus?: string }).importStatus === "pending_ai_interpretation",
+      );
+      if (
+        (result as { importStatus?: string }).importStatus !== "pending_ai_extraction" &&
+        (result as { importStatus?: string }).importStatus !== "pending_ai_interpretation"
+      ) {
+        // Sync upload path (no polling): polls do not run, so fetch status for year-mismatch / inferredYearFromFile
+        getImportStatus({ cityId, inventoryId, importedFileId: result.id })
+          .unwrap()
+          .then(setLastImportStatus)
+          .catch((err) =>
+            logger.debug(
+              { err, cityId, inventoryId, importedFileId: result.id },
+              "Import status fetch after sync upload failed",
+            ),
+          );
+        setTimeout(() => goToNextStep(), 150);
+      }
+    } catch (error: unknown) {
+      const message = getApiErrorMessage(error, t("failed-to-upload-file"));
+      makeErrorToast(t("upload-failed"), message);
+    }
+  };
+
+  const handleContinue = () => {
+    if (activeStep < importStepCount - 1) {
+      // Small delay for smooth transition
+      setTimeout(() => {
+        goToNextStep();
+      }, 80);
+    }
+  };
+
+  // Reset scroll when moving between import segments (combined stepper UX).
+  useEffect(() => {
+    window.scrollTo({ top: 0, behavior: "smooth" });
+  }, [activeStep]);
+
+  const handleRemoveFile = () => {
+    stopExtractionPolling();
+    stopInterpretPolling();
+    stopUploadPolling();
+    setIsExtractInProgress(false);
+    setIsInterpretInProgress(false);
+    setUploadedFile(null);
+    setImportedFileId(null);
+    setLastImportStatus(null);
+    setPdfPendingExtraction(false);
+    setTabularPendingInterpretation(false);
+    setStep(0);
+    fileYearMismatchToastShownRef.current = false;
+  };
+
+  const handleExtractWithAi = async () => {
+    if (!importedFileId || !inventoryId) return;
+    stopExtractionPolling();
+    setIsExtractInProgress(true);
+    setExtractionProgress(null);
+    startExtractionPolling();
+    try {
+      const result = await extractImport({
+        cityId,
+        inventoryId,
+        importedFileId,
+      }).unwrap();
+      if ("accepted" in result && result.accepted) return;
+      stopExtractionPolling();
+      setIsExtractInProgress(false);
+      setPdfPendingExtraction(false);
+      await getImportStatus({
+        cityId,
+        inventoryId,
+        importedFileId,
+      }).unwrap();
+      setTimeout(() => goToNextStep(), 150);
+    } catch (error: unknown) {
+      stopExtractionPolling();
+      setIsExtractInProgress(false);
+      setExtractionProgress(null);
+      const apiMessage = getApiErrorMessage(error);
+      const message =
+        apiMessage === "Inventory not found for the target year"
+          ? t("inventory-not-found-for-target-year")
+          : apiMessage || t("ai-extraction-failed-default");
+      makeErrorToast(t("extraction-failed"), message);
+    }
+  };
+
+  const handleInterpretWithAi = async () => {
+    if (!importedFileId || !inventoryId) return;
+    stopInterpretPolling();
+    setIsInterpretInProgress(true);
+    setExtractionProgress(null);
+    makeInfoToast(
+      t("interpreting-file"),
+      t("interpreting-file-description"),
+    );
+    try {
+      const result = await interpretImport({
+        cityId,
+        inventoryId,
+        importedFileId,
+      }).unwrap();
+      if ("accepted" in result && result.accepted) {
+        startInterpretPolling();
+        return;
+      }
+      setIsInterpretInProgress(false);
+      setExtractionProgress(null);
+      if ((result as { importStatus?: string }).importStatus === "failed") {
+        const errorLog = (result as { errorLog?: string | null }).errorLog;
+        makeErrorToast(
+          t("interpretation-failed") ?? "Interpretation failed",
+          resolveErrorMessage(errorLog, t("ai-extraction-failed-default"), t),
+        );
+        return;
+      }
+      setTabularPendingInterpretation(false);
+      setTimeout(() => goToNextStep(), 150);
+    } catch (error: unknown) {
+      setIsInterpretInProgress(false);
+      setExtractionProgress(null);
+      const message = getApiErrorMessage(
+        error,
+        t("ai-extraction-failed-default"),
+      );
+      makeErrorToast(t("interpretation-failed") ?? "Interpretation failed", message);
+    }
+  };
+
+  // Handle beforeunload event (browser refresh/close)
+  useEffect(() => {
+    const handleBeforeUnload = (e: BeforeUnloadEvent) => {
+      if (hasUnsavedProgress) {
+        e.preventDefault();
+        // Modern browsers ignore custom messages, but we still need to call preventDefault
+        e.returnValue = "";
+        return "";
+      }
+    };
+
+    if (hasUnsavedProgress) {
+      window.addEventListener("beforeunload", handleBeforeUnload);
+      return () => window.removeEventListener("beforeunload", handleBeforeUnload);
+    }
+  }, [hasUnsavedProgress]);
+
+  // Handle route changes
+  useEffect(() => {
+    prevPathnameRef.current = pathname;
+  }, [pathname]);
+
+  // Handle navigation attempts (back button, router.push, etc.)
+  const handleNavigation = (navigationFn: () => void) => {
+    if (hasUnsavedProgress) {
+      setPendingNavigation(() => navigationFn);
+      setShowDataLossModal(true);
+    } else {
+      navigationFn();
+    }
+  };
+
+  const handleConfirmLeave = () => {
+    setShowDataLossModal(false);
+    if (pendingNavigation) {
+      pendingNavigation();
+      setPendingNavigation(null);
+    }
+  };
+
+  const handleCancelLeave = () => {
+    setShowDataLossModal(false);
+    setPendingNavigation(null);
+  };
+
+  return (
+    <>
+      <Box pt={16} pb={16} maxW="full" mx="auto" w="1090px">
+        <Button
+          variant="ghost"
+          onClick={() => {
+            if (activeStep === 0) {
+              handleNavigation(() => {
+                if (onExitFirstStep) {
+                  onExitFirstStep();
+                } else {
+                  router.back();
+                }
+              });
+            } else {
+              goToPrevStep();
+            }
+          }}
+          pl={0}
+          color="content.link"
+        >
+          <Icon as={MdArrowBack} boxSize={6} />
+          {t("go-back")}
+        </Button>
+        <Box
+          display="flex"
+          flexDirection={{ base: "column", md: "row" }}
+          columnGap={{ md: "48px" }}
+          rowGap={{ base: "48px", md: "0px" }}
+          alignItems="flex-start"
+          mt={{ base: 8, md: 16 }}
+          mb={48}
+          w={"1090px"}
+          mx="auto"
+          position="relative"
+          minH="400px"
+        >
+          <Box w="full" position="relative" minH="400px" overflow="hidden">
+            <AnimatePresence mode="wait">
+              {activeStep === 0 && (
+                <motion.div
+                  key="step-0"
+                  initial={{ opacity: 0, x: 100 }}
+                  animate={{ opacity: 1, x: 0 }}
+                  exit={{ opacity: 0, x: -100 }}
+                  transition={{ duration: 0.2, ease: "easeInOut" }}
+                >
+                  <Box w="full" display="flex" flexDirection="column" gap="24px">
+                    {inventoryHasData && (
+                      <HStack
+                        gap={3}
+                        px={4}
+                        py={3}
+                        bg="orange.50"
+                        border="1px solid"
+                        borderColor="orange.200"
+                        borderRadius="md"
+                        align="flex-start"
+                      >
+                        <Icon as={MdWarning} boxSize={5} color="orange.500" mt="2px" />
+                        <Box>
+                          <Text fontWeight="semibold" color="orange.800" fontSize="sm">
+                            {t("inventory-already-has-data-title")}
+                          </Text>
+                          <Text color="orange.700" fontSize="sm" mt={1}>
+                            {t("inventory-already-has-data-description", {
+                              year: inventoryYear,
+                            })}
+                          </Text>
+                        </Box>
+                      </HStack>
+                    )}
+                    <UploadFileStep
+                      t={t}
+                      cityName={inventory?.city?.name}
+                      uploadedFile={uploadedFile}
+                      onFileUpload={inventoryHasData ? () => { } : handleFileUpload}
+                      onRemoveFile={handleRemoveFile}
+                      isUploading={isUploadingFile || isUploadPolling}
+                    />
+                    {pdfPendingExtraction && (isExtracting || isExtractInProgress) && (
+                      <Box w="full" mt={2}>
+                        <Text fontSize="sm" color="content.tertiary" mb={2}>
+                          {extractionProgress && extractionProgress.total > 1
+                            ? t("extracting-chunk-progress", {
+                              current: extractionProgress.current,
+                              total: extractionProgress.total,
+                            })
+                            : t("breaking-into-chunks")}
+                        </Text>
+                        {extractionProgress && extractionProgress.total > 1 ? (
+                          <Box
+                            w="full"
+                            h="8px"
+                            bg="background.subtle"
+                            borderRadius="10px"
+                            overflow="hidden"
+                          >
+                            <Box
+                              h="full"
+                              bg="interactive.primary"
+                              borderRadius="10px"
+                              transition="width 0.3s ease"
+                              w={`${(extractionProgress.current / extractionProgress.total) * 100}%`}
+                            />
+                          </Box>
+                        ) : (
+                          <Box
+                            w="full"
+                            h="8px"
+                            bg="background.subtle"
+                            borderRadius="10px"
+                            overflow="hidden"
+                            position="relative"
+                          >
+                            <motion.div
+                              style={{
+                                position: "absolute",
+                                left: 0,
+                                top: 0,
+                                height: "100%",
+                                width: "40%",
+                              }}
+                              animate={{ x: ["0%", "250%"] }}
+                              transition={{
+                                duration: 1.5,
+                                repeat: Infinity,
+                                ease: "easeInOut",
+                              }}
+                            >
+                              <Box
+                                h="full"
+                                w="full"
+                                bg="interactive.primary"
+                                borderRadius="10px"
+                              />
+                            </motion.div>
+                          </Box>
+                        )}
+                      </Box>
+                    )}
+                    {tabularPendingInterpretation && (isInterpreting || isInterpretInProgress) && (
+                      <Box w="full" mt={2}>
+                        <Text fontSize="sm" color="content.tertiary" mb={2}>
+                          {extractionProgress && extractionProgress.total > 1
+                            ? t("interpreting-chunk-progress", {
+                              current: extractionProgress.current,
+                              total: extractionProgress.total,
+                            })
+                            : t("interpreting-file-description")}
+                        </Text>
+                        {extractionProgress && extractionProgress.total > 1 ? (
+                          <Box
+                            w="full"
+                            h="8px"
+                            bg="background.subtle"
+                            borderRadius="10px"
+                            overflow="hidden"
+                          >
+                            <Box
+                              h="full"
+                              bg="interactive.primary"
+                              borderRadius="10px"
+                              transition="width 0.3s ease"
+                              w={`${(extractionProgress.current / extractionProgress.total) * 100}%`}
+                            />
+                          </Box>
+                        ) : (
+                          <Box
+                            w="full"
+                            h="8px"
+                            bg="background.subtle"
+                            borderRadius="10px"
+                            overflow="hidden"
+                            position="relative"
+                          >
+                            <motion.div
+                              style={{
+                                position: "absolute",
+                                left: 0,
+                                top: 0,
+                                height: "100%",
+                                width: "40%",
+                              }}
+                              animate={{ x: ["0%", "250%"] }}
+                              transition={{
+                                duration: 1.5,
+                                repeat: Infinity,
+                                ease: "easeInOut",
+                              }}
+                            >
+                              <Box
+                                h="full"
+                                w="full"
+                                bg="interactive.primary"
+                                borderRadius="10px"
+                              />
+                            </motion.div>
+                          </Box>
+                        )}
+                      </Box>
+                    )}
+                  </Box>
+                </motion.div>
+              )}
+              {activeStep === 1 && importedFileId && inventoryId && (
+                <motion.div
+                  key="step-1"
+                  initial={{ opacity: 0, x: 100 }}
+                  animate={{ opacity: 1, x: 0 }}
+                  exit={{ opacity: 0, x: -100 }}
+                  transition={{ duration: 0.2, ease: "easeInOut" }}
+                >
+                  <InventoryMappingStep
+                    t={t}
+                    cityId={cityId}
+                    cityName={inventory?.city?.name}
+                    inventoryId={inventoryId}
+                    importedFileId={importedFileId}
+                    mappingOverrides={mappingOverrides}
+                    onMappingChange={(columnName, mappedKey) => {
+                      setMappingOverrides((prev) => ({
+                        ...prev,
+                        [columnName]: mappedKey,
+                      }));
+                    }}
+                    canContinue={canContinueMapping}
+                  />
+                </motion.div>
+              )}
+              {activeStep === 2 && importedFileId && inventoryId && (
+                <motion.div
+                  key="step-2"
+                  initial={{ opacity: 0, x: 100 }}
+                  animate={{ opacity: 1, x: 0 }}
+                  exit={{ opacity: 0, x: -100 }}
+                  transition={{ duration: 0.2, ease: "easeInOut" }}
+                >
+                  {isImportInProgress ? (
+                    <ImportingInventoryStep
+                      t={t}
+                      cityName={inventory?.city?.name}
+                    />
+                  ) : (
+                    <ReviewConfirmStep
+                      t={t}
+                      cityId={cityId}
+                      cityName={inventory?.city?.name}
+                      inventoryId={inventoryId}
+                      importedFileId={importedFileId}
+                      onEditMapping={goToPrevStep}
+                    />
+                  )}
+                </motion.div>
+              )}
+            </AnimatePresence>
+          </Box>
+        </Box>
+        <Box
+          bg="white"
+          w="full"
+          position="fixed"
+          bottom={0}
+          left={0}
+          pb={8}
+          px={1}
+          zIndex={9999}
+          transition="all"
+          data-import-bottom-bar
+        >
+          <Box w="full" display="flex" flexDir="column" gap="32px">
+            <Box w="full">
+              <Box w="full">
+                <ProgressSteps steps={steps} currentStep={progressCurrentStep} />
+              </Box>
+            </Box>
+            <Box w="full" display="flex" justifyContent="end" px="135px">
+              {activeStep === 0 && (
+                <Button
+                  w="auto"
+                  gap="8px"
+                  py="16px"
+                  px="24px"
+                  onClick={
+                    pdfPendingExtraction
+                      ? handleExtractWithAi
+                      : tabularPendingInterpretation
+                        ? handleInterpretWithAi
+                        : handleContinue
+                  }
+                  h="64px"
+                  disabled={
+                    inventoryHasData ||
+                    !uploadedFile ||
+                    !importedFileId ||
+                    (pdfPendingExtraction && (isExtracting || isExtractInProgress)) ||
+                    (tabularPendingInterpretation && (isInterpreting || isInterpretInProgress))
+                  }
+                  loading={
+                    (pdfPendingExtraction && (isExtracting || isExtractInProgress)) ||
+                    (tabularPendingInterpretation && (isInterpreting || isInterpretInProgress))
+                  }
+                >
+                  <Text
+                    fontFamily="button.md"
+                    fontWeight="600"
+                    letterSpacing="wider"
+                  >
+                    {pdfPendingExtraction || tabularPendingInterpretation
+                      ? t("extract-with-ai")
+                      : t("continue")}
+                  </Text>
+                  <MdArrowForward height="24px" width="24px" />
+                </Button>
+              )}
+              {activeStep === 1 && (
+                <Button
+                  w="auto"
+                  gap="8px"
+                  py="16px"
+                  px="24px"
+                  onClick={handleContinue}
+                  h="64px"
+                  disabled={fileYearMismatch || !canContinueMapping}
+                >
+                  <Text
+                    fontFamily="button.md"
+                    fontWeight="600"
+                    letterSpacing="wider"
+                  >
+                    {t("continue")}
+                  </Text>
+                  <MdArrowForward height="24px" width="24px" />
+                </Button>
+              )}
+              {activeStep === 2 && importedFileId && inventoryId && (
+                <HStack gap="16px">
+                  <Button
+                    variant="outline"
+                    minW="171px"
+                    gap="8px"
+                    py="16px"
+                    px="24px"
+                    h="64px"
+                    disabled={isImportInProgress}
+                    onClick={() => {
+                      handleNavigation(() => {
+                        setUploadedFile(null);
+                        setImportedFileId(null);
+                        setLastImportStatus(null);
+                        setPdfPendingExtraction(false);
+                        setTabularPendingInterpretation(false);
+                        setIsExtractInProgress(false);
+                        setIsInterpretInProgress(false);
+                        setIsImportInProgress(false);
+                        setMappingOverrides({});
+                        setExtractionProgress(null);
+                        setStep(0);
+                      });
+                    }}
+                  >
+                    <Text
+                      fontFamily="button.md"
+                      fontWeight="600"
+                      letterSpacing="wider"
+                    >
+                      {t("cancel")}
+                    </Text>
+                  </Button>
+                  <ImportButton
+                    cityId={cityId}
+                    inventoryId={inventoryId}
+                    importedFileId={importedFileId}
+                    mappingOverrides={mappingOverrides}
+                    onImportProgressChange={setIsImportInProgress}
+                    onImport={() => {
+                      onComplete(inventoryId);
+                    }}
+                    t={t}
+                  />
+                </HStack>
+              )}
+            </Box>
+          </Box>
+        </Box>
+      </Box>
+      <DataLossWarningModal
+        isOpen={showDataLossModal}
+        onOpenChange={setShowDataLossModal}
+        onConfirm={handleConfirmLeave}
+        onCancel={handleCancelLeave}
+        t={t}
+      />
+    </>
+  );
+}
